Moved parseLayoutValue to MapUtils
